feat:add DeleteProfilePhoto.tsx, add internationality for add and delete profile photo. refactor: errorText.
