+++ src/shared/locales/en.ts
@@ -116,5 +116,18 @@ export const en = {
     favorites: 'Favorites',
     log_out: 'Log Out',
   },
+  add_profile_photo: {
+    add_profile_photo_text: 'Add a Profile Photo',
+    text_of_button_select_from_comp: 'Select from Computer',
+    error_typy_of_photo: 'Error! The format of the uploaded photo must be PNG and JPEG.',
+    error_size_photo: 'Error! Photo size must be less than 10 MB!',
+    save_button: ' Save',
+  },
+  delete_photo_of_profile: {
+    title_of_modal: 'Delete Photo',
+    text: 'Are you sure you want to delete the photo?',
+    button_yes: 'Yes',
+    button_no: 'No',
+  },
 }
 export type LangType = typeof en
