+++ src/widgets/addProfilePhoto/AddProfilePhotoModal.module.scss
@@ -1,0 +1,64 @@
+
+.errorText{
+  width: 100%;
+  margin-bottom: 30px;
+  padding: 6px 24px;
+
+  font-size: var(--font-size-s);
+  font-weight: var(--font-weight-400);
+line-height: var(--line-height-m);
+  text-align: center;
+
+  background-color: var(--color-danger-900);
+  border:1px solid var(--color-danger-500);
+  border-radius: 2px
+}
+
+.defaultProfilePhotoBlock {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 222px;
+  height: 228px;
+  margin-bottom: 60px;
+
+  background-color: var(--color-dark-500);
+  border-radius: 2px;
+}
+
+.cropContainer{
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  .profilePhotoBlock {
+    position: relative;
+
+    width: 332px;
+    height: 340px;
+    margin-bottom: 36px;
+
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: cover;
+
+
+  }
+
+
+}
+
+.buttonBox{
+  display: flex;
+  justify-content: flex-end;
+}
+
+.box {
+  display: flex;
+  flex-direction: column;
+  align-items:center;
+  justify-content: center;
+
+  margin-bottom: 84px;
+}
