+++ src/widgets/addProfilePhoto/ProfilePhotoForGeneralInfo.tsx
