+++ src/widgets/addProfilePhoto/index.ts
@@ -1,0 +1,1 @@
+export * from './ProfilePhotoForGeneralInfo'
