+++ src/widgets/myProfile/MyProfile.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react'
 import s from './MyProfile.module.scss'
 
 import { useTranslation } from '@/shared/lib'
-import { ProfilePhotoForGeneralInfo } from '@/widgets/profilePhoto/ProfilePhotoForGeneralInfo'
+import { ProfilePhotoForGeneralInfo } from '@/widgets/addProfilePhoto/ProfilePhotoForGeneralInfo'
 type Props = {}
 export const MyProfile: FC<Props> = () => {
   return (
