--- src/widgets/profilePhoto/index.ts
@@ -1,1 +1,0 @@
-export * from 'src/widgets/profilePhoto/ProfilePhotoForGeneralInfo'
